Use one shared null and bash program
